+++ src/pin_layout.rs
@@ -184,14 +184,14 @@ fn get_pin_widget(
             let pullup_pick = pullup_picklist(pull, board_pin_number, bcm_pin_number.unwrap());
             if is_left {
                 Row::new()
-                    .push(led(16.0, 16.0, pin_state.level))
-                    .push(pin_state.view())
+                    .push(led(16.0, 16.0, pin_state.get_level()))
+                    .push(pin_state.chart())
                     .push(pullup_pick)
             } else {
                 Row::new()
                     .push(pullup_pick)
-                    .push(led(16.0, 16.0, pin_state.level))
-                    .push(pin_state.view())
+                    .push(led(16.0, 16.0, pin_state.get_level()))
+                    .push(pin_state.chart())
             }
         }
